fix RT#81332 744aaba059 bloats the B compilers
reverts 744aaba059, which moved the B bootstrap'ing from run-time to
BEGIN, which is a big problem in all B compilers, since any compiler
does not know if a B constant is needed inside the users code or just
imported by using the compiler.  E.g. every compiled .plc bytecode
file is now 13K bigger, all B constants are imported now. Similar for
the other compilers.

The only reason of the move was to avoid to stub B::OPf_KIDS.

Note: The test is still wrong. I can only repro it in the B::Bytecode
compiler so far.