shaped arrays: rename from sized arrays
Rationale: Discussion with #perl6. They named it shaped because this fits
better for multidimensial arrays.  We will eventually add multidim also, and
it is less confusing for the casual user of cperl and perl6 then.
Even if "shaped" reminds me on "shaved", and "sized" is easier to pronounce
and remember. "shaped" are only structs, but not arrays. Oh well.
if multi, they are multi-sized.

Add more samples to the feature docs, but it should be in perldata.

shaped arrays: add unchecked op variants as _u

harmonize the names for the unchecked variants for all types.
use the _u suffix.
implemented for native and boxed typed, but not yet
used (enabled in the compiler)

shaped arrays: fix init_shaped, add t/op/aelem_u.t

just fixed the native type check.
scalar does not work yet.

not native tests yet, as the compiler does not add the box ops, and uses
the new ops yet.
even the current tests still use the old aelemfast_lex ops, not the
new typed ones.

shaped arrays: compile-time check out of bounds

add ck_aelem, add SHAPED flag to sv_dump.
check shaped array indices at compile-time.

TODO: promote the aelem ops if shaped or typed.

shaped array: compile-time check invalid ops

Throw "Invalid modification of shaped array: op var"

Detect shape modifying ops: shift, unshift, pop, push.
but for unshift, push only the 0 and 1 arg cases yet.
splice is only checked at run-time, as we would need to check many
const args.
Disable the temp. SvREADONLY so we can assign now.

do not initialize the shaped array with undef, as it is readonly.
rather use the magic NULL value.

add run-time errors for the compile-time counterparts.
Note that the var name is not yet fixed there.

enable ck_pad for padav also, coming only from newAVREF, via PADANY.
but the op is not chained yet.

compile to aelemfast_lex_u.
flip negative constant indices to positive and allow max 255,
not just 127.

on lval return an empty sv body to be filled in as in aelemfast,
which is just an optimized padav. needed e.g. for sassign lval.
This is the inlined behaviour of aelemfast_lex, with ignoring magic,
av_fetch and bounds checks.

aelem_u: add OPpMAYBE_LVSUB to op_private
this private bit is allowed, the others not.