[CHANGE] % with one negative integer argument
Describe how cperl behaves differently to perl5 with both integer
arguments with one negative to %. It behaves like with use integer.
With a typed integer variable or a integer constant cperl uses i_modulo
due to type dispatch. This behaves now like with the libc, and not like
perl5 and is consistent with constant folding being done on promoted ops.

Also provide rationale for all other incompatible changes.