SVf_PROTECT: replace by explicit checks
in Internals::SvREADONLY and  SvREADONLY_off().
We need the last remaining SV bit for something more important.
CORE does not violate SvREADONLY_off for our protected values.
With extensions check explicitly.

One CHANGE:
Internals::SvREADONLY( <protect value>, 0) croaks immediately,
and not just when trying to assign a value later to it.
This is arguably the better behaviour to avoid magic errors at
a distance.